Return dynamically-calculated atom type populations from Configuration using a KeyedVector.

+++ src/classes/species_atomic.cpp
@@ -270,9 +270,9 @@ double Species::mass() const
 }
 
 // Calculate and return atom type populations
-KeyedVector<const AtomType *, double> Species::atomTypePopulations() const
+KeyedVector<const AtomType *, int> Species::atomTypePopulations() const
 {
-    KeyedVector<const AtomType *, double> result;
+    KeyedVector<const AtomType *, int> result;
     for (const auto &i : atoms_)
         if (i.atomType())
             result.add(i.atomType().get(), 1);
